+++ cd-extras/private/Core.ps1
@@ -112,7 +112,7 @@ function RegisterCompletions([string[]] $commands, $param, $target) {
 }
 
 function WriteLog($message) {
-  if ((Get-Variable cde) -and ($cde | Get-Member -MemberType ScriptMethod _logger)) { 
+  if ((Get-Variable cde) -and ($cde | Get-Member -MemberType ScriptMethod _logger)) {
     $cde._logger($message)
   }
   else { Write-Verbose $message }