+++ tests/cd-extras.Tests.ps1
@@ -33,31 +33,41 @@ Describe 'cd-extras' {
   Describe 'cd' {
     It 'works with spaces' {
       cd 'powershell/directory with spaces'
-      CurrentDir | Should Be 'directory with spaces'
+      CurrentDir | Should -Be 'directory with spaces'
     }
 
     It 'pushes to the undo stack' {
-      (Get-Stack -Undo) | should be $null
+      (Get-Stack -Undo) | Should -Be $null
       cd powershell
       cd src
-      (Get-Stack -Undo).Count | should be 2
+      (Get-Stack -Undo).Count | Should -Be 2
       (Get-Stack -Undo).Name | select -First 1 | should -Be 'powershell'
     }
 
     It 'does not push duplicates' {
-      (Get-Stack -Undo) | should be $null
+      (Get-Stack -Undo) | Should -Be $null
       cd powershell
       cd src
       cd ../src
-      (Get-Stack -Undo).Count | should be 2
-      (Get-Stack -Undo).Name | select -First 1 | should -Be 'powershell'
+      (Get-Stack -Undo).Count | Should -Be 2
+      (Get-Stack -Undo).Name | Select -First 1 | Should -Be 'powershell'
     }
 
     It 'supports piping values' {
       @('powershell', 'src') | cd
-      CurrentDir | Should Be 'src'
+      CurrentDir | Should -Be 'src'
       cd-
-      CurrentDir | Should Not Be 'src'
+      CurrentDir | Should -Not -Be 'src'
+    }
+
+    It 'supports literal paths' {
+      cd 'pow/directory[with]squarebrackets'
+      CurrentDir | Should -Be 'directory[with]squarebrackets'
+    }
+
+    It 'is not fazed by a trailing separator' {
+      cd 'pow/directory[with]squarebrackets/'
+      CurrentDir | Should -Be 'directory[with]squarebrackets'
     }
   }
 
@@ -65,37 +75,37 @@ Describe 'cd-extras' {
     It 'moves back to previous directory' {
       cd powershell; cd src
       cd-
-      CurrentDir | Should Be powershell
+      CurrentDir | Should -Be powershell
     }
 
     It 'moves back "n" locations' {
       cd powershell; cd src; cd Modules; cd Shared
       cd- 2
-      CurrentDir | Should Be src
+      CurrentDir | Should -Be src
     }
 
     It 'works with the zsh style syntax' {
       cd powershell; cd src; cd Modules; cd Shared
       cd -2
-      CurrentDir | Should Be src
+      CurrentDir | Should -Be src
     }
 
     It 'works with the tilde syntax' {
       cd powershell; cd src; cd Modules; cd Shared
       cd ~2
-      CurrentDir | Should Be src
+      CurrentDir | Should -Be src
     }
 
     It 'moves back to a named location' {
       cd powershell; cd src; cd Modules; cd Shared
       cd- src
-      CurrentDir | Should Be src
+      CurrentDir | Should -Be src
     }
 
     It 'prefers an exact match if available' {
       cd powershell; cd src; cd libpsl-native; cd test; cd googletest
       cd- TestDrive:${/}powershell${/}src
-      CurrentDir | Should Be src
+      CurrentDir | Should -Be src
     }
 
     It 'throws if the named location cannot be found' {
@@ -106,25 +116,25 @@ Describe 'cd-extras' {
     It 'matches more than one segment if necessary' {
       cd powershell; cd src; cd Modules; cd Shared
       cd- src/mod
-      CurrentDir | Should Be Modules
+      CurrentDir | Should -Be Modules
     }
 
     It 'pops a directory with literal square brackets' {
       cd 'powershell/directory`[with`]squarebrackets/one'; cd ..
       cd-
-      CurrentDir | Should Be one
+      CurrentDir | Should -Be one
     }
 
     It 'pushes current directory when moving into a directory with literal square brackets' {
       cd powershell; cd 'directory`[with`]squarebrackets/one'
       cd-
-      CurrentDir | Should Be powershell
+      CurrentDir | Should -Be powershell
     }
 
     It 'pushes current directory when CDing into a directory with question mark' {
       cd powershell; cd demos/A?ure;
       cd-
-      CurrentDir | Should Be powershell
+      CurrentDir | Should -Be powershell
     }
   }
 
@@ -132,31 +142,31 @@ Describe 'cd-extras' {
     It 'moves forward on the stack' {
       cd powershell; cd src; cd-
       cd+
-      CurrentDir | Should Be src
+      CurrentDir | Should -Be src
     }
 
     It 'moves forward "n" locations' {
       cd powershell; cd src; cd Modules; cd Shared; cd-; cd-
       cd+ 2
-      CurrentDir | Should Be Shared
+      CurrentDir | Should -Be Shared
     }
 
     It 'works with the zsh style syntax' {
       cd powershell; cd src; cd Modules; cd Shared; cd-; cd-
       cd +2
-      CurrentDir | Should Be Shared
+      CurrentDir | Should -Be Shared
     }
 
     It 'works with the tilde syntax' {
       cd powershell; cd src; cd Modules; cd Shared; cd-; cd-
       cd ~~2
-      CurrentDir | Should Be Shared
+      CurrentDir | Should -Be Shared
     }
 
     It 'moves forward to a named location' {
       cd powershell; cd src; cd Modules; cd Shared; cd-; cd-
       cd+ shared
-      CurrentDir | Should Be Shared
+      CurrentDir | Should -Be Shared
     }
 
     It 'throws if the named location cannot be found' {
@@ -168,7 +178,7 @@ Describe 'cd-extras' {
     It 'pops a directory with square brackets' {
       cd 'powershell/directory`[with`]squarebrackets/one'; cd-
       cd+
-      CurrentDir | Should Be one
+      CurrentDir | Should -Be one
     }
   }
 
@@ -177,9 +187,9 @@ Describe 'cd-extras' {
       cd ./powershell/src/Modules
       cd ../../demos/Apache
       cdb
-      CurrentDir | Should Be Modules
+      CurrentDir | Should -Be Modules
       cdb
-      CurrentDir | Should Be Apache
+      CurrentDir | Should -Be Apache
     }
   }
 
@@ -187,43 +197,43 @@ Describe 'cd-extras' {
     It 'can move up two directories' {
       Set-Location ./powershell/src/Modules/Shared/Microsoft.PowerShell.Utility
       cd ...
-      CurrentDir | Should Be Modules
+      CurrentDir | Should -Be Modules
     }
 
     It 'can move up three directories' {
       Set-Location ./powershell/src/Modules/Shared/Microsoft.PowerShell.Utility
       cd ....
-      CurrentDir | Should Be src
+      CurrentDir | Should -Be src
     }
 
     It 'works even when CD_PATH is set' {
       setocd CD_PATH @('TestDrive:\powershell\src\')
       Set-Location ./powershell/src/Modules/Shared/Microsoft.PowerShell.Utility
       cd ...
-      CurrentDir | Should Be Modules
+      CurrentDir | Should -Be Modules
     }
   }
 
   Describe 'Path shortening cd' {
     It 'can move up multiple directories' {
       cd ./pow/src/Mod
-      CurrentDir | Should Be Modules
+      CurrentDir | Should -Be Modules
     }
 
     It 'works with explicit Path parameter' {
       cd -Path ./pow/src/Mod
-      CurrentDir | Should Be Modules
+      CurrentDir | Should -Be Modules
     }
 
     It 'works in the registry provider' -Skip:(!$IsWindows) {
       cd HKLM:/
       cd so/mic
-      CurrentDir | Should Be Microsoft
+      CurrentDir | Should -Be Microsoft
     }
 
     It 'supports the double dot operator' {
       cd pow/src/Typ..Gen
-      CurrentDir | Should Be TypeCatalogGen
+      CurrentDir | Should -Be TypeCatalogGen
     }
   }
 
@@ -231,25 +241,25 @@ Describe 'cd-extras' {
     It 'can be called explicitly' {
       Set-Location .\powershell\src\Modules\Shared\Microsoft.PowerShell.Utility
       cd: shared Unix
-      Get-Location | Split-Path | Should Be TestDrive:${/}powershell${/}src${/}Modules${/}Unix
+      Get-Location | Split-Path | Should -Be TestDrive:${/}powershell${/}src${/}Modules${/}Unix
     }
 
     It 'can replace more than one path segment' {
       Set-Location .\powershell\demos\Apache\Apache
       cd: Apache/Apache crontab/CronTab
-      CurrentDir | Should Be CronTab
+      CurrentDir | Should -Be CronTab
     }
 
     It 'works with two arg cd' {
       Set-Location .\powershell\src\Modules\Shared\Microsoft.PowerShell.Utility
       cd Shared Unix
-      Get-Location | Split-Path | Should Be TestDrive:${/}powershell${/}src${/}Modules${/}Unix
+      Get-Location | Split-Path | Should -Be TestDrive:${/}powershell${/}src${/}Modules${/}Unix
     }
 
     It 'leaves an entry on the Undo stack' {
       Set-Location .\powershell\src\Modules\Shared\Microsoft.PowerShell.Utility
       cd Shared Unix
-      (Get-Stack -Undo).Name | Select -First 1 | Should Be Microsoft.PowerShell.Utility
+      (Get-Stack -Undo).Name | Select -First 1 | Should -Be Microsoft.PowerShell.Utility
     }
 
     It 'throws if the replaceable text is not in the current directory name' {
@@ -267,49 +277,49 @@ Describe 'cd-extras' {
     It 'moves up one level if no arguments given' {
       Set-Location p*\src\Sys*\Format*\common\Utilities
       Step-Up
-      CurrentDir | Should Be common
+      CurrentDir | Should -Be common
     }
 
     It 'can navigate upward by a given number of directories' {
       Set-Location p*\src\Sys*\Format*\common\Utilities
       Step-Up 4
-      CurrentDir | Should Be src
+      CurrentDir | Should -Be src
     }
 
     It 'can navigate upward by name' {
       Set-Location p*\src\Sys*\Format*\common\Utilities
       Step-Up src
-      CurrentDir | Should Be src
+      CurrentDir | Should -Be src
     }
 
     It 'can navigate upward by partial name' {
       Set-Location p*\src\Sys*\Format*\common\Utilities
       Step-Up com
-      CurrentDir | Should Be common
+      CurrentDir | Should -Be common
     }
 
     It 'can navigate within the registry on Windows' -Skip:(!$IsWindows) {
       Set-Location HKLM:\Software\Microsoft\Windows\CurrentVersion
       Step-Up 2
-      CurrentDir | Should Be Microsoft
+      CurrentDir | Should -Be Microsoft
     }
 
     It 'can navigate within the registry on Windows by name' -Skip:(!$IsWindows) {
       Set-Location HKLM:\Software\Microsoft\Windows\CurrentVersion
       Step-Up mic
-      CurrentDir | Should Be Microsoft
+      CurrentDir | Should -Be Microsoft
     }
 
     It 'can navigate by full name if no matching leaf name' {
       Set-Location powershell\src\Modules\Shared\
       Step-Up (Resolve-Path ..).Path
-      CurrentDir | Should Be Modules
+      CurrentDir | Should -Be Modules
     }
 
     It 'does not choke on root directory full path' {
       Set-Location $PSScriptRoot
       Step-Up (Get-Location).Drive.Root
-      $PWD | Should Be (Get-Location).Drive.Root
+      $PWD | Should -Be (Get-Location).Drive.Root
     }
 
     It 'throws if the given name part is not found' {
@@ -321,21 +331,21 @@ Describe 'cd-extras' {
   Describe 'Get-Up' {
     It 'returns the parent directory by default' {
       Set-Location pow*/docs/git
-      Get-Up | Should Be (Resolve-Path ..).Path
+      Get-Up | Should -Be (Resolve-Path ..).Path
     }
 
     It 'can take an arbitrary path' {
-      Get-Up -From powershell\docs\git | Should be (Resolve-Path powershell\docs).Path
+      Get-Up -From powershell\docs\git | Should -Be (Resolve-Path powershell\docs).Path
     }
 
     It 'return $null when n is out of range' {
       Set-Location pow*/docs/git
-      Get-Up 255 | Should BeNullOrEmpty
+      Get-Up 255 | Should -BeNullOrEmpty
     }
 
     It 'is fine with square brackets' {
       Set-Location 'powershell/directory`[with`]squarebrackets/one'
-      Get-Up | Split-Path -Leaf | Should Be 'directory[with]squarebrackets'
+      Get-Up | Split-Path -Leaf | Should -Be 'directory[with]squarebrackets'
     }
 
     It 'should return -From when n is 0' {
@@ -347,19 +357,19 @@ Describe 'cd-extras' {
     It 'exports ancestors when Export switch set' {
       Set-Location p*\src\Sys*\Format*\common\Utilities
       Get-Ancestors -Export -Force
-      $Global:common | Should Be (Resolve-Path ..).Path
-      $Global:formatAndOutput | Should Be (Resolve-Path ../..).Path
+      $Global:common | Should -Be (Resolve-Path ..).Path
+      $Global:formatAndOutput | Should -Be (Resolve-Path ../..).Path
       # ...
     }
 
     It 'does not choke on duplicate directory names' {
       Set-Location powershell/powershell
       $xup = (Get-Ancestors).Path
-      $xup[0] | Should BeLike ($pwd.Path | Split-Path)
-      $xup[1] | Should BeLike ($pwd.Path | Split-Path | Split-Path)
+      $xup[0] | Should -BeLike ($pwd.Path | Split-Path)
+      $xup[1] | Should -BeLike ($pwd.Path | Split-Path | Split-Path)
     }
 
-    It 'should not export the root directory when switch set' {
+    It 'should -not export the root directory when switch set' {
       $xup = Get-Ancestors -From ~ -ExcludeRoot
       $xup.Path | Should -Not -Contain (Resolve-Path ~).Drive.Root
     }
@@ -380,62 +390,62 @@ Describe 'cd-extras' {
     It 'can change directory' {
       Set-Location powershell
       src
-      CurrentDir | Should Be src
+      CurrentDir | Should -Be src
     }
 
     It 'can change directory using a partial match' {
       Set-Location powershell
       sr
-      CurrentDir | Should Be src
+      CurrentDir | Should -Be src
     }
 
     It 'can change directory using multiple partial path segments' {
       Set-Location powershell
       sr/Res
-      CurrentDir | Should Be ResGen
+      CurrentDir | Should -Be ResGen
     }
 
     It 'can navigate up multiple levels' {
       Set-Location p*\src\Sys*\Format*\common\Utilities
       .....
-      CurrentDir | Should Be src
+      CurrentDir | Should -Be src
     }
 
     It 'does nothing if more than one word given' {
       Set-Location powershell
       sr x
-      CurrentDir | Should Be powershell
+      CurrentDir | Should -Be powershell
     }
 
     It 'does nothing when turned off' {
       Set-CdExtrasOption -Option AUTO_CD -Value $false
       Set-Location powershell
       { src } | Should Throw
-      CurrentDir | Should Be powershell
+      CurrentDir | Should -Be powershell
     }
 
     It 'supports the double dot operator' {
       pow/src/Typ..Gen
-      CurrentDir | Should Be TypeCatalogGen
+      CurrentDir | Should -Be TypeCatalogGen
     }
 
     It 'works in the registry provider' -Skip:(!$IsWindows) {
       cd HKLM:
       so/mic
-      CurrentDir | Should Be Microsoft
+      CurrentDir | Should -Be Microsoft
     }
 
     It 'supports tilde undo syntax' {
       cd powershell; cd src; cd Modules; cd Shared
       ~2
-      CurrentDir | Should Be src
+      CurrentDir | Should -Be src
     }
 
     It 'supports tilde redo syntax' {
       cd powershell; cd src; cd Modules; cd Shared
       cd- 2
       ~~2
-      CurrentDir | Should Be Shared
+      CurrentDir | Should -Be Shared
     }
   }
 
@@ -444,7 +454,7 @@ Describe 'cd-extras' {
       $Global:psh = Resolve-Path ./pow*/src/Mod*/Shared/*.Host
       Set-CdExtrasOption CDABLE_VARS $true
       cd psh
-      CurrentDir | Should Be 'Microsoft.PowerShell.Host'
+      CurrentDir | Should -Be 'Microsoft.PowerShell.Host'
     }
 
     It 'works with AUTO_CD' {
@@ -454,7 +464,7 @@ Describe 'cd-extras' {
       $Global:psh = Resolve-Path ./pow*/src/Mod*/Shared/*.Host
       $Global:__cdeUnderTest = $true
       psh
-      CurrentDir | Should Be 'Microsoft.PowerShell.Host'
+      CurrentDir | Should -Be 'Microsoft.PowerShell.Host'
     }
   }
 
@@ -462,14 +472,14 @@ Describe 'cd-extras' {
     It 'moves to the expected location' {
       $cde.NOARG_CD = '/'
       cd
-      (Get-Location).Path | Should Be (Resolve-Path /).Path
+      (Get-Location).Path | Should -Be (Resolve-Path /).Path
     }
 
     It 'leaves an entry in the Undo stack' {
       $startLocation = (Get-Location).Path
       $cde.NOARG_CD = '~'
       cd
-      (Get-Stack -Undo).Path | select -First 1 | Should Be $startLocation
+      (Get-Stack -Undo).Path | select -First 1 | Should -Be $startLocation
     }
 
     It 'does not change location when null' {
@@ -486,13 +496,13 @@ Describe 'cd-extras' {
         'TestDrive:\powershell\src\', 'TestDrive:\powershell\docs\')
         
       cd ResGen
-      CurrentDir | Should Be resgen
+      CurrentDir | Should -Be resgen
     }
 
     It 'works when there is one exact match and several partial matches' {
       Set-CdExtrasOption -Option CD_PATH -Value @('powershell\src\Modules\')
       cd Windows
-      CurrentDir | Should Be windows
+      CurrentDir | Should -Be windows
     }
 
     It 'does not search CD_PATH when given directory is rooted or relative' {
@@ -504,17 +514,17 @@ Describe 'cd-extras' {
   Describe 'Expand-Path' {
     It 'returns expected expansion Windows style' {
       Expand-Path p/s/m/U |
-        Should Be (Join-Path $TestDrive powershell\src\Modules\Unix)
+      Should -Be (Join-Path $TestDrive powershell\src\Modules\Unix)
     }
 
     It 'expands $PWD if given an empty string' {
-      # last result should be powershell/powershell
-      Expand-Path '' | Select -last 1 -Expand Name | Should Be 'powershell'
+      # last result Should -Be powershell/powershell
+      Expand-Path '' | Select -last 1 -Expand Name | Should -Be 'powershell'
     }
 
     It 'returns expected expansion relative style' {
       Expand-Path ./p/s/m/U |
-        Should Be (Join-Path $TestDrive powershell\src\Modules\Unix)
+      Should -Be (Join-Path $TestDrive powershell\src\Modules\Unix)
     }
 
     It 'expands rooted paths' -Skip:(!$IsWindows) {
@@ -523,7 +533,7 @@ Describe 'cd-extras' {
     }
 
     It 'can return multiple expansions' {
-      (Expand-Path ./p/s/m/s/M).Length | Should Be 2
+      (Expand-Path ./p/s/m/s/M).Length | Should -Be 2
     }
 
     It 'considers CD_PATH for expansion' {
@@ -533,16 +543,16 @@ Describe 'cd-extras' {
 
     It 'expands around periods' {
       Expand-Path p/s/.Con |
-        Should Be (Join-Path $TestDrive powershell\src\Microsoft.PowerShell.ConsoleHost)
+      Should -Be (Join-Path $TestDrive powershell\src\Microsoft.PowerShell.ConsoleHost)
     }
 
     It 'supports the double dot operator' {
       Expand-Path pow/src/Typ..Gen |
-        Should Be (Join-Path $TestDrive powershell\src\TypeCatalogGen)
+      Should -Be (Join-Path $TestDrive powershell\src\TypeCatalogGen)
     }
 
     It 'works in Windows registry' -Skip:(!$IsWindows) {
-      (Expand-Path HKLM:\Soft\Mic\).Count | Should BeGreaterOrEqual 1
+      (Expand-Path HKLM:\Soft\Mic\).Count | Should -BeGreaterOrEqual 1
     }
   }
 
@@ -552,7 +562,7 @@ Describe 'cd-extras' {
     }
 
     It 'shows the redo and undo stacks' {
-      (Get-Stack).Count | Should Be 2
+      (Get-Stack).Count | Should -Be 2
     }
 
     It 'shows the undo stack' {
@@ -590,17 +600,17 @@ Describe 'cd-extras' {
   Describe 'Clear-Stack' {
     It 'clears the undo stack' {
       cd powershell
-      Get-Stack -Undo | Should Not BeNullOrEmpty
+      Get-Stack -Undo | Should -Not -BeNullOrEmpty
       Clear-Stack -Undo
-      Get-Stack -Undo | Should BeNullOrEmpty
+      Get-Stack -Undo | Should -BeNullOrEmpty
     }
 
     It 'clears the redo stack' {
       cd powershell
       cd-
-      Get-Stack -Redo | Should Not BeNullOrEmpty
+      Get-Stack -Redo | Should -Not -BeNullOrEmpty
       Clear-Stack -Redo
-      Get-Stack -Redo | Should BeNullOrEmpty
+      Get-Stack -Redo | Should -BeNullOrEmpty
     }
   }
 
@@ -612,7 +622,7 @@ Describe 'cd-extras' {
         $actual | Should HaveCount 3
 
         function ShouldContain($likeStr) {
-          $actual | Where-Object { $_ -like $likeStr } | Should Not BeNullOrEmpty
+          $actual | Where-Object { $_ -like $likeStr } | Should -Not -BeNullOrEmpty
         }
 
         ShouldContain "*test${/}csharp${/}"
@@ -622,23 +632,23 @@ Describe 'cd-extras' {
 
       It 'expands around periods' {
         $actual = CompletePaths -wordToComplete './pow/s/.SDK'
-        $actual.CompletionText | Should BeLike "*powershell${/}src${/}Microsoft.PowerShell.SDK${/}"
+        $actual.CompletionText | Should -BeLike "*powershell${/}src${/}Microsoft.PowerShell.SDK${/}"
       }
 
       It 'completes directories with spaces correctly' {
         $actual = CompletePaths  -wordToComplete 'pow/directory with spaces/child one'
-        $actual.CompletionText | Should BeLike "'*${/}child one${/}'"
+        $actual.CompletionText | Should -BeLike "'*${/}child one${/}'"
       }
 
       It 'completes relative directories with spaces correctly' {
         $actual = CompletePaths -wordToComplete './pow/directory with spaces/child one'
-        $actual.CompletionText | Should BeLike "'*${/}child one${/}'"
+        $actual.CompletionText | Should -BeLike "'*${/}child one${/}'"
       }
 
       It 'completes relative directories with a relative prefix' {
         Set-Location $PSScriptRoot
         $actual = CompletePaths -wordToComplete '../cd-extras/public'
-        $actual.CompletionText | Should Be "..${/}cd-extras${/}public${/}"
+        $actual.CompletionText | Should -Be "..${/}cd-extras${/}public${/}"
       }
 
       It 'completes absolute paths with an absolute prefix' {
@@ -650,45 +660,45 @@ Describe 'cd-extras' {
 
       It 'expands multiple dots' {
         Set-Location p*\src\Sys*\Format*\common\Utilities
-        (CompletePaths -wordToComplete '...').CompletionText | Should Match 'FormatAndOutput'
+        (CompletePaths -wordToComplete '...').CompletionText | Should -Match 'FormatAndOutput'
       }
 
       It 'completes CDABLE_VARS' {
         setocd CDABLE_VARS
         $Global:dir = Resolve-Path ./powershell/src
-        (CompletePaths -wordToComplete 'dir').CompletionText | Should Match 'src'
+        (CompletePaths -wordToComplete 'dir').CompletionText | Should -Match 'src'
       }
 
       It 'completes file paths' {
         Set-Location $PSScriptRoot
         (CompletePaths -filesOnly -wordToComplete './samp').CompletionText |
-          Should Match "sampleStructure.txt"
+        Should -Match "sampleStructure.txt"
       }
 
       It 'provides usable registry paths' -Skip:(!$IsWindows) {
         (CompletePaths -dirsOnly -wordToComplete 'HKLM:\Soft\Mic').CompletionText |
-          Should Match "HKLM:\\Software\\Microsoft"
+        Should -Match "HKLM:\\Software\\Microsoft"
       }
 
-      It 'escapes square brackets' {
+      It 'completes paths with square brackets' {
         $actual = CompletePaths -wordToComplete 'pow/directory[with]squarebrackets/o'
-        $actual.CompletionText | Should BeLike "'*${/}powershell${/}directory*squarebrackets${/}one${/}'"
+        $actual.CompletionText | Should -BeLike "'*${/}powershell${/}directory*squarebrackets${/}one${/}'"
       }
 
       It 'appends a directory separator given a single dot' {
         $actual = CompletePaths -wordToComplete '.'
-        @($actual)[0].CompletionText | Should Be ".${/}"
+        @($actual)[0].CompletionText | Should -Be ".${/}"
       }
 
       It 'supports the double dot operator' {
         $actual = CompletePaths -wordToComplete 'pow/src/Typ..Gen'
-        $actual.CompletionText | Should BeLike "*src${/}TypeCatalogGen${/}"
+        $actual.CompletionText | Should -BeLike "*src${/}TypeCatalogGen${/}"
       }
 
       It 'truncates long menu items' {
         $actual = CompletePaths -wordToComplete 'pow/reallyreally..long'
         $actual.ListItemText.Length | Should -BeLessThan ($actual.CompletionText | Split-Path -Leaf).Length
-        $actual.ListItemText.Length | Should Be $cde.MaxMenuLength
+        $actual.ListItemText.Length | Should -Be $cde.MaxMenuLength
       }
 
       It 'colourises output only if option set' {
@@ -712,6 +722,13 @@ Describe 'cd-extras' {
         $actual = CompletePaths -wordToComplete 'zzzzzzzzz'
         $actual | Should -Be $null
       }
+
+      It 'uses the force switch when the command being completed does not have one' {
+        $git = Get-Item powershell/.git/
+        $git.Attributes = "Hidden"
+        $actual = CompletePaths -wordToComplete 'pow/.git' -commandName 'Get-Date' | Select -Expand ListItemText
+        $actual | Should -Be @('.git', '.github')
+      }
     }
 
     Describe 'Stack expansion' {
@@ -719,7 +736,7 @@ Describe 'cd-extras' {
         Set-LocationEx powershell
         Set-LocationEx src
         $actual = CompleteStack -wordToComplete '' -commandName 'Undo'
-        $actual.Count | Should BeGreaterThan 1
+        $actual.Count | Should -BeGreaterThan 1
       }
 
       It 'expands the redo stack' {
@@ -727,7 +744,7 @@ Describe 'cd-extras' {
         Set-LocationEx src
         cd- 2
         $actual = CompleteStack -wordToComplete '' -commandName 'Redo'
-        $actual.Count | Should BeGreaterThan 1
+        $actual.Count | Should -BeGreaterThan 1
       }
 
       It 'uses index completion when menu completion is on' {
@@ -735,7 +752,7 @@ Describe 'cd-extras' {
         Set-LocationEx src
         $cde.MenuCompletion = $true
         $actual = CompleteStack -wordToComplete '' -commandName 'Undo'
-        $actual[0].CompletionText | Should Be 1
+        $actual[0].CompletionText | Should -Be 1
       }
 
       It 'uses the full path when menu completion is off' {
@@ -743,14 +760,14 @@ Describe 'cd-extras' {
         Set-LocationEx src
         $cde.MenuCompletion = $false
         $actual = CompleteStack -wordToComplete '' -commandName 'Undo'
-        $actual[0].CompletionText | Should BeLike "TestDrive:${/}powershell"
+        $actual[0].CompletionText | Should -BeLike "TestDrive:${/}powershell"
       }
 
       It 'uses the full path when only one completion is available' {
         Set-LocationEx powershell
         $cde.MenuCompletion = $true
         $actual = CompleteStack -wordToComplete '' -commandName 'Undo'
-        $actual[0].CompletionText | Should BeLike "testdrive:${/}"
+        $actual[0].CompletionText | Should -BeLike "testdrive:${/}"
       }
 
       It 'returns $null result if no completions available' {
@@ -764,35 +781,35 @@ Describe 'cd-extras' {
       It 'expands ancestors' {
         Set-Location p*\src\Sys*\Format*\common\Utilities
         $actual = CompleteAncestors -wordToComplete ''
-        $actual.Count | Should BeGreaterThan 5
+        $actual.Count | Should -BeGreaterThan 5
       }
 
       It 'uses index completion when menu completion is on' {
         Set-Location ./powershell/demos/Apache
         $cde.MenuCompletion = $true
         $actual = CompleteAncestors -wordToComplete ''
-        $actual[0].CompletionText | Should Be 1
+        $actual[0].CompletionText | Should -Be 1
       }
 
       It 'uses the full path when menu completion is off' {
         Set-Location ./powershell/demos/Apache
         $cde.MenuCompletion = $false
         $actual = CompleteAncestors -wordToComplete ''
-        $actual[0].CompletionText | Should BeLike "*powershell${/}demos"
+        $actual[0].CompletionText | Should -BeLike "*powershell${/}demos"
       }
 
       It 'can complete against a more than one path segment' {
         Set-Location ./powershell/demos/Apache
         $actual = CompleteAncestors -wordToComplete 'll/de'
         $actual | Should -HaveCount 1
-        $actual[0].CompletionText | Should BeLike "*powershell${/}demos"
+        $actual[0].CompletionText | Should -BeLike "*powershell${/}demos"
       }
 
       It 'can match against a previously completed full path' {
         Set-Location ./powershell/demos/Apache
         $target = CompleteAncestors -wordToComplete 'demos'
         $actual = CompleteAncestors -wordToComplete $target[0].CompletionText
-        $actual[0].CompletionText | Should BeLike $target[0].CompletionText
+        $actual[0].CompletionText | Should -BeLike $target[0].CompletionText
       }
 
       It 'returns $null result if no completions available' {